// flyteconsole/packages/zapp/console/src/components/Launch/LaunchForm/types.ts
export interface InputTypeDefinition {
  literalType: LiteralType;
  type: InputType;
  subtype?: InputTypeDefinition;
  listOfSubTypes?: InputTypeDefinition[];
}

export interface BlobValue {
  dimensionality: BlobDimensionality | string;
  format?: string;
  uri: string;
}

export interface UnionValue {
  value: InputValue;
  typeDefinition: InputTypeDefinition;
}

export interface NoneValue {}

export type InputValue = string | number | boolean | Date | BlobValue | UnionValue | NoneValue;